GS: clarify/correct some logic: should be if platform = A30, not if platform != Flip.

+++ spruce/scripts/gameswitcher.sh
@@ -116,8 +116,7 @@ done <$LIST_FILE
 
 # send signal USR2 to joystickinput to switch to KEYBOARD MODE
 # this allows joystick to be used as DPAD in game switcher
-if [ ! "$PLATFORM" = "Flip" ]; then
-    killall -q -USR2 joystickinput
+[ "$PLATFORM" = "A30" ] && killall -q -USR2 joystickinput
 fi
 # launch the switcher program
 # Usage: switcher image_list title_list [-s speed] [-b on|off] [-m on|off] [-t on|off] [-ts speed] [-n on|off] [-d command]
@@ -181,9 +180,8 @@ while : ; do
 done
 
 # send signal USR1 to joystickinput to switch to ANALOG MODE
-if [ ! "$PLATFORM" = "Flip" ]; then
-    killall -q -USR1 joystickinput
-fi
+[ "$PLATFORM" = "A30" ] && killall -q -USR1 joystickinput
+
 # launch game with return index
 if [ $RETURN_INDEX -gt 0 ]; then
     # get command that launches the game
